+++ mods/common/src/schemas/index.ts
@@ -25,6 +25,9 @@ export {
   deleteApplicationDocumentSchema,
   getApplicationDocumentSchema,
   getApplicationEvidenceSchema,
+  setApplicationMapUrlSchema,
+  listEvidenceQueueSchema,
+  getEvidenceTaskSchema,
   approveApplicationSchema,
   rejectApplicationSchema,
   promoteApplicationSchema,
@@ -64,6 +67,8 @@ export {
   type DeleteApplicationDocumentInput,
   type GetApplicationDocumentInput,
   type GetApplicationEvidenceInput,
+  type SetApplicationMapUrlInput,
+  type GetEvidenceTaskInput,
   type ApproveApplicationInput,
   type RejectApplicationInput,
   type PromoteApplicationInput,
@@ -92,6 +97,7 @@ export {
   DEFAULT_MIN_BUSINESS_PHOTOS,
   evaluateTransition,
   evidenceStatus,
+  evidenceProgress,
   transitionBlockCode,
   type ReviewAction,
   type RejectionReason,
@@ -462,3 +468,4 @@ export {
   type TaskView,
   type TaskFiringView
 } from "./task.js";
+export { isMapUrl, buildMapUrl, MAX_MAP_URL_LENGTH } from "./mapUrl.js";